feat(suite): add password manager as an experimental feature
