+++ packages/suite/src/views/password-manager/PasswordManager/Tag.tsx
