+++ packages/suite/src/views/password-manager/index.tsx
@@ -1,0 +1,7 @@
+import { useLayout } from 'src/hooks/suite';
+import { PasswordManager } from './PasswordManager/PasswordManager';
+
+export default () => {
+    useLayout('Password manager');
+    return <PasswordManager />;
+};
